Created new Metadata provider for TheMovieDB: MovieDbEpisodeProvider
Most of the required logic already existed in
MovieDbEpisodeImageProvider.
In order to avoid duplicated code, I have moved all common code to a new
base class "MovieDbProviderBase" from which MovieDbEpisodeProvider and
MovieDbEpisodeImageProvider are inheriting now.
